Main Window Screen Directives
Commit changes the way new screens are loaded in the main window. Before
this commit, sidebar and titlebar were present in index.html file, and
new screen (i.e., Home, Settings, etc.) was loaded as an ngView in
index.html using angular routing. This method has a major disadvantage,
which is index.html will always contain the window wrapper code, which
makes it unusable for creating new windows like qpopup, etc.

To resolve this problem, a main_window view is being used, with screens
as directives that can be loaded using 'changescreen' directive in
sidebar item links. On click, 'changescreen' directive clears out the
html of '#gf-screen-container' element using jqLite and appends new
screen directive based on which item is clicked in the sidebar